--- BirtherDay/Common/Components/BDButton.swift
@@ -1,8 +1,0 @@
-//
-//  BDButton.swift
-//  BirtherDay
-//
-//  Created by Rama on 5/28/25.
-//
-
-import Foundation
